% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_transects.R
\name{create_transects}
\alias{create_transects}
\title{Create transect}
\usage{
create_transects(
  shape_obj,
  design = "systematic",
  design.angle = 0,
  line.length,
  segmentize = FALSE,
  length_segs,
  crs = 2154,
  ...
)
}
\arguments{
\item{shape_obj}{sf object. The shape of the study site.}

\item{design}{character. Variable describing the type of design. Either "random", "systematic", "eszigzag" (equal-spaced zigzag), "eszigzagcom" (equal spaced zigzag with complementary lines) or "segmentedgrid". See dssd package for more information.}

\item{design.angle}{numeric. Value detailing the angle of the design. A value of -1 will cause a random design angle to be generated. See dssd package for more information. Bydefault = 0.}

\item{line.length}{numeric. The approximative total line length desired (in m).}

\item{segmentize}{boolean. TRUE = transects are segmentized and a seg_ID is given to each segment. FALSE = transects are not segmentized. By default = FALSE.}

\item{length_segs}{numeric. Length of the segments desired.}

\item{crs}{numeric. Projection system. By default = 2154.}

\item{...}{All other arguments that could be used in the make.design function. See dssd package for more information.}
}
\value{
sf object. The created transects.
}
\description{
Create transect
}
\examples{

library(ggplot2)
data("shape_courseulles")

# ------------------------------
# Example 1 : systematic parallel transects with a approximative total length of 400000m

transects <- create_transects(shape_obj = shape_courseulles,
                             design = "systematic",
                             line.length = 400000)
# Plot
ggplot() +
      geom_sf(data = shape_courseulles) +
      geom_sf(data = transects) 


# ------------------------------
# Example 2 : zigzag transects with a approximative total length of 400000m 
# that are segmentized with a length of 2000m per segment
transects <- create_transects(shape_obj = shape_courseulles,
                             design = "eszigzag",
                             line.length = 400000,
                             design.angle = 90,
                             segmentize = TRUE,
                             length_segs = 2000)

# Plot
pal <- rep(c("#98C1D9","#EE6C4D","#293241"),nrow(transects)) 
ggplot() +
      geom_sf(data = shape_courseulles) +
      geom_sf(data = transects, aes(colour = seg_ID))+
      scale_colour_manual(values=pal) +
      theme(legend.position = "none")



# ------------------------------
# Example 3 : systematic parallel transects with a approximative total length of 400000m
# that are segmentized with a length of 2000m per segment 
# and crop according to the map_obj grid containing densities

# Create map 
map <- create_density_map(shape_obj = shape_courseulles,
                              N = 200,
                              grid_size = 1000,
                              density_type = "gradient",
                              gradient_direction = "N",
                              wavelength = 20000,
                              amplitude = 500)

# Create transects
transects <- create_transects(shape_obj = map,
                             design = "systematic",
                             line.length = 400000,
                             design.angle = 2,
                             segmentize = TRUE,
                             length_segs = 2000)

# Plot
pal <- rep(c("#98C1D9","#EE6C4D","#293241"),nrow(transects)) 
ggplot() +
      geom_sf(data = map, color = "#D9D9D9") +
      geom_sf(data = transects, aes(colour = seg_ID))+
      scale_colour_manual(values = pal) +
      theme(legend.position = "none")

}
