% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depth_courseulles.R
\docType{data}
\name{depth_courseulles}
\alias{depth_courseulles}
\title{This data contains the depth in the Courseulles study area (crs = 2154)}
\description{
This data contains the depth in the Courseulles study area (crs = 2154)
}
\author{
Maud Queroue \email{maud.queroue@gmail.com}
}
\keyword{Courseulles}
\keyword{data}
\keyword{depth}
