% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_obs.R
\name{esw_hn}
\alias{esw_hn}
\title{esw_hn}
\usage{
esw_hn(sigma)
}
\arguments{
\item{sigma}{numeric. Sigma of half normal function.}
}
\value{
numeric. esw value associated to the sigma given.
}
\description{
esw_hn
}
\keyword{internal}
