% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_density.R
\name{simulate_density}
\alias{simulate_density}
\title{Simulate density maps}
\usage{
simulate_density(
  shape_obj,
  grid_size = 1000,
  density_type,
  gradient_direction,
  wavelength,
  amplitude,
  nb_hotspots,
  crs = 2154
)
}
\arguments{
\item{shape_obj}{sf object. The shape of the study site.}

\item{grid_size}{numeric. Length of length of square side. By default: 1000.}

\item{density_type}{character. 'uniform", 'gradient', 'random', 'covariate'.}

\item{gradient_direction}{character. Only for "gradient" \code{density_type}. Where the highest density comes from. 'N' for North,'NE','E'... and 'C" for center}

\item{wavelength}{numeric. Only for "gradient" and "random" \code{density_type}. Wavelength of the hotspots of density created. Maximum wavelength for "random" \code{density_type}.}

\item{amplitude}{numeric. Only for "gradient" and "random" \code{density_type}. Amplitude of the hotspots of density created. Maximum amplitude for "random" \code{density_type}.}

\item{nb_hotspots}{numeric. Only for "random" \code{density_type}. Number of random hotspots to be created.}

\item{crs}{numeric. Only for "random" \code{density_type}. Number of random hotspots to be created. By default: 2154.}
}
\value{
The map in a sf dataframe contaning the simulated density. The dataframe contains two columns, \code{density} simulated for each square of the grid and \code{geometry} of the grid.
}
\description{
\code{simulate_density} allows to simulate density with different spatial variation.
}
\details{
This function use create the \code{make.density} and the \code{add.hotspot} functions of the \code{dsims} package.

To simulate density on the map, a grid is created. The length of square side could be chosen by the user with \code{grid_size}.

If \code{density_type} is '\code{gradient}', one hotspot is added on one direction \code{gradient_direction} (North, North-East, East, South-East, South, South-West, West, North-West and center). The hotspot added have a \code{amplitude} and a \code{wavelength} chosen by the user.

If \code{density_type} is '\code{random}', several hotspots, the number is chosen by the user with the argument \code{nb_hotspots}, are added randomly in the area.The amplitude and wavelength of the differents hotspot are chosen randomly between a minimum of 1 and a maximum corresponding to the \code{amplitude} and \code{wavelength} chosen by the user.
}
\examples{

library(ggplot2)
data("shape_courseulles")

# ------------------------------
# Example 1: Simulate a map with a gradient density from the North
#-------------------------------

map <- simulate_density(shape_obj = shape_courseulles,
                              density_type = "gradient",
                              gradient_direction = "N",
                              wavelength = 20000,
                              amplitude = 10)

# Plot
ggplot() +
  geom_sf(data = map, aes(fill = density)) +
  theme(panel.background = element_rect(fill = "white"),
        panel.grid.major = element_line(colour = "#EDEDE9"))


# ------------------------------
# Example 2: Simulate a map with a random density
# ------------------------------

map <- simulate_density(shape_obj = shape_courseulles,
                              density_type = "random",
                              wavelength = 10000,
                              amplitude = 15,
                              nb_hotspots = 10
                              
)

# Plot 
ggplot() +
  geom_sf(data = map, aes(fill = density)) +
  theme(panel.background = element_rect(fill = "white"),
        panel.grid.major = element_line(colour = "#EDEDE9"))

}
