% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_ind.R
\name{simulate_ind}
\alias{simulate_ind}
\title{Simulate individuals}
\usage{
simulate_ind(map_obj, N, crs = 2154)
}
\arguments{
\item{map_obj}{sf dataframe. Sf map with a colum containg density informations}

\item{N}{numeric. The number of individuals desired in the area.}

\item{crs}{numeric. Projection system. By default: 2154.}
}
\value{
The simulated individuals in a dataframe. The dataframe contains three columns, \code{x} longitude of the simulated individual, \code{y} latitude of the simulated individual and \code{size} the number of individuals simulated on the location (the possibility of simulate groupe will comming).
}
\description{
\code{simulate_ind} allows to simulate a approximate number of individuals with a inhomogenous Poisson point process according to the densities provided.
}
\details{
This function need a grid sf object with density assosciated to each square of the grid.
For each simulated individuals, a size is associated. For now, it is always one, but in the future it will be possible to simulate groups.
}
\examples{

library(ggplot2)
data("shape_courseulles")

# First, create a map with a gradient density from the North with 500 individuals in the area
map <- simulate_density(shape_obj = shape_courseulles,
                              grid_size = 1000,
                              density_type = "gradient",
                              gradient_direction = "N",
                              wavelength = 20000,
                              amplitude = 100)

# Then simulate the presence of individuals in the study area 
ind <- simulate_ind(map_obj = map, N = 200)

# Plot
ggplot() +
    geom_sf(data = map, aes(fill = density)) +
    geom_point(data = ind, aes(x = x, y = y))
  
}
