% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_obs.R
\name{simulate_obs}
\alias{simulate_obs}
\title{Detection function}
\usage{
simulate_obs(
  ind_obj,
  transect_obj,
  key,
  esw = NA,
  g_zero = 1,
  truncation = NA,
  crs = 2154
)
}
\arguments{
\item{ind_obj}{dataframe. Individuals simulated with their coordinates.}

\item{transect_obj}{sf dataframe. Transect/segments data.}

\item{key}{character. Form of the detection function "hn" or "unif".}

\item{esw}{numeric. Effective strip width in m. This argument is needed only for the 'hn' form.}

\item{g_zero}{numeric. The detection probability at distance 0. By default 1.}

\item{truncation}{numeric. A single numeric value in m describing the longest distance at which an object may be observed. By default = NA.}

\item{crs}{numeric. Projection system. By default = 2154.}
}
\value{
dataframe. The dist dataframe with new columns : "proba" (numeric values between 0 and 1), the probability of being detection and 'detected' (0 or 1)informing if the individual is detected by the sample design.
}
\description{
Detection function
}
\examples{

library(ggplot2)
data("shape_courseulles")

# First, create a density map
map <- simulate_density(shape_obj = shape_courseulles,
                          density_type = "gradient",
                          gradient_direction = "N",
                          wavelength = 20000,
                          amplitude = 500
                          
)

# Then simulate the presence of individuals in the study area 
ind <- simulate_ind(map_obj = map, N = 500)

# Then create transects
transects <- simulate_transects(shape_obj = map,
                             design = "systematic",
                             line.length = 400000,
                             design.angle = 2,
                             segmentize = TRUE,
                             length_segs = 2000)

# Finally, detection of individuals simulated on the map 
# according to simulated densities and simulated transects :

# ------------------------------
# Example 1 : With a half normal detection probability equal to 1 at 0 distance (g_zero)
# with a effective half width of 180 m.
# ------------------------------

obs <- simulate_obs(ind_obj = ind,
                 transect_obj = transects,
                 key = "hn",
                 g_zero = 1,
                 esw = 180)

# Plot detection probability
ggplot(obs, aes(x=distance, y=proba)) +
  geom_point(color = "#EE6C4D") +
  xlim(0,500) 

# Plot on map
ggplot() +
  geom_sf(data = map, fill = "#CDDAFD", color = "#CDDAFD") +
  geom_sf(data = transects, color = "black") +
  geom_point(data=obs[obs$detected==0,], aes(x=x, y=y), shape=20, color="#051923") +
  geom_point(data=obs[obs$detected==1,], aes(x=x, y=y), shape=21, fill="#EE6C4D") +
  labs(caption = paste("Sightings = ", sum(obs$detected), sep = " ")) +
  theme(legend.position = "none",
        panel.background = element_rect(fill = "white"),
        panel.grid.major = element_line(colour = "#EDEDE9"))


# ------------------------------
# Example 2 : With a uniforme detection probability equal to 1 below distance 200 m.
# ------------------------------

obs <- simulate_obs(ind_obj = ind,
                 transect_obj = transects,
                 key = "unif",
                 g_zero = 1,
                 truncation = 200)

# Plot detection probability
ggplot(obs, aes(x=distance, y=proba)) +
  geom_point(color = "#EE6C4D") +
  xlim(0,500) 

# Plot on map
ggplot() +
  geom_sf(data = map, fill = "#CDDAFD", color = "#CDDAFD") +
  geom_sf(data = transects, color = "black") +
  geom_point(data=obs[obs$detected==0,], aes(x=x, y=y), shape=20, color="#051923") +
  geom_point(data=obs[obs$detected==1,], aes(x=x, y=y), shape=21, fill="#EE6C4D") +
  labs(caption = paste("Sightings = ", sum(obs$detected), sep = " ")) +
  theme(legend.position = "none",
        panel.background = element_rect(fill = "white"),
        panel.grid.major = element_line(colour = "#EDEDE9"))

}
