% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_transects.R
\name{simulate_transects}
\alias{simulate_transects}
\title{Simulate Transects}
\usage{
simulate_transects(
  shape_obj,
  design = "systematic",
  design.angle = 0,
  line.length,
  segmentize = FALSE,
  length_segs,
  crs = 2154,
  ...
)
}
\arguments{
\item{shape_obj}{sf object. The shape of the study site or a density grid.}

\item{design}{character. Variable describing the type of design. Either \code{random}, \code{systematic} (parallel), \code{eszigzag} (equal-spaced zigzag), \code{eszigzagcom} (equal spaced crossed zigzag). See \code{dssd} package for more information.}

\item{design.angle}{numeric. Value detailing the angle of the design. A value of -1 will cause a random design angle to be generated. See \code{dssd} package for more information. By default: 0.}

\item{line.length}{numeric. The approximative total line length desired in m.}

\item{segmentize}{boolean. If \code{TRUE} transects are segmentized and a \code{seg_ID} is given to each segment. If \code{FALSE}transects are not segmentized. By default: \code{FALSE}.}

\item{length_segs}{numeric. Length of the segments desired for segmentation.}

\item{crs}{numeric. Projection system. By default: 2154.}

\item{...}{All other arguments that could be used in the make.design function. See dssd package for more information.}
}
\value{
The simulated transects corresponding to the differents conditions in a sf dataframe. The dataframe contains four columns: \code{transect} number, \code{seg_ID} (a unique identifiant for each segments/transect), \code{effort} (length of each segment/transect) and \code{geometry}.
}
\description{
\code{simulate_transects} allows to simulate transects with different designs and total length. It also allows to segmentize the simulated transects.
}
\details{
This function use create the \code{make.design} function of the \code{dssd} package.

If segmentize is \code{TRUE}, \code{seg_ID} is noted "1-2" for transect 1 segment 2. If segmentize is \code{FALSE}, \code{seg_ID} is identical to the transect number (first column).
}
\examples{

library(ggplot2)
data("shape_courseulles")

# ------------------------------
# Example 1 : parallel transects with a approximative total length of 400000m
# ------------------------------

transects <- simulate_transects(shape_obj = shape_courseulles,
                             design = "systematic",
                             line.length = 400000)
# Plot
ggplot() +
      geom_sf(data = shape_courseulles, fill = "#EDEDE9") +
      geom_sf(data = transects) +
      theme(panel.background = element_rect(fill = "white"),
      panel.grid.major = element_line(colour = "#EDEDE9"))

# ------------------------------
# Example 2 : zigzag transects with a approximative total length of 400000m 
# that are segmentized with a length of approximately 2000m per segment
# ------------------------------

transects <- simulate_transects(shape_obj = shape_courseulles,
                             design = "eszigzag",
                             line.length = 400000,
                             design.angle = 90,
                             segmentize = TRUE,
                             length_segs = 2000)

# Plot
pal <- rep(c("#98C1D9","#EE6C4D","#293241"),nrow(transects)) 
ggplot() +
      geom_sf(data = shape_courseulles, fill = "#EDEDE9") +
      geom_sf(data = transects, aes(colour = seg_ID))+
      scale_colour_manual(values=pal) +
      theme(legend.position = "none",
      panel.background = element_rect(fill = "white"),
      panel.grid.major = element_line(colour = "#EDEDE9"))


# ------------------------------
# Example 3 : systematic parallel transects with a approximative total length of 400000m
# that are segmentized with a length of approximately 2000m per segment 
# the `shape_obj` is a density map simulated with `simulate_density`
#-------------------------------

# Create map 
map <- simulate_density(shape_obj = shape_courseulles,
                              grid_size = 1000,
                              density_type = "gradient",
                              gradient_direction = "N",
                              wavelength = 20000,
                              amplitude = 500)

# Create transects
transects <- simulate_transects(shape_obj = map,
                             design = "systematic",
                             line.length = 400000,
                             design.angle = 2,
                             segmentize = TRUE,
                             length_segs = 2000)

# Plot
pal <- rep(c("#98C1D9","#EE6C4D","#293241"),nrow(transects)) 
ggplot() +
      geom_sf(data = map, color = "#D9D9D9") +
      geom_sf(data = transects, aes(colour = seg_ID))+
      scale_colour_manual(values = pal) +
      theme(legend.position = "none",
      panel.background = element_rect(fill = "white"),
      panel.grid.major = element_line(colour = "#EDEDE9"))

}
