% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape_courseulles.R
\docType{data}
\name{shape_courseulles}
\alias{shape_courseulles}
\title{This data contains the shape of the Courseulles study area}
\description{
This data contains the shape of the Courseulles study area
}
\author{
Maud Quéroué \email{maud.queroue@gmail.com}
}
\keyword{Courseulles}
\keyword{data}
\keyword{shape}
