% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_ind.R
\name{simulate_ind}
\alias{simulate_ind}
\title{Simulate individuals with a inhomogenous Poisson point process}
\usage{
simulate_ind(map_obj, crs = 2154)
}
\arguments{
\item{map_obj}{sf dataframe. Sf map with a colum containg density informations}

\item{crs}{numeric. Projection system. By default = 2154.}
}
\value{
Dataframe. Les localisations des individus.
}
\description{
Simulate individuals with a inhomogenous Poisson point process
}
\examples{

library(ggplot2)
data("shape_courseulles")

# First, create a map with a gradient density from the North with 500 individuals in the area
map <- create_density_map(shape_obj = shape_courseulles,
                              N = 200,
                              grid_size = 1000,
                              density_type = "gradient",
                              gradient_direction = "N",
                              wavelength = 20000,
                              amplitude = 500
                              
)


# Then simulate the presence of individuals in the study area 
ind <- simulate_ind(map_obj = map)


# Plot
ggplot() +
    geom_sf(data = map, aes(fill = density)) +
    geom_point(data = ind, aes(x = x, y = y))
  
}
