% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_ind.R
\name{simulate_ind}
\alias{simulate_ind}
\title{Simulate individuals with a inhomogenous Poisson point process}
\usage{
simulate_ind(map_obj)
}
\arguments{
\item{map_obj}{Dataframe. Sf map with a colum containg density informations density_m}
}
\value{
Dataframe. Les localisations des individus.
}
\description{
Simulate individuals with a inhomogenous Poisson point process
}
\examples{

library(ggplot2)
data("shape_courseulles")

# First, create a map with a gradient density from the North with 500 individuals in the area
map_obj <- create_density_map(shape_obj = shape_courseulles,
                              N = 200,
                              grid_size = 1000,
                              density_type = "gradient",
                              gradient_direction = "N",
                              wavelength = 30000,
                              amplitude = 70
                              
)


# Then simulate the presence of individuals in the study area with a inhomogenous Poisson point process
obs_obj <- simulate_ind(map_obj = map_obj)


# Plot results
ggplot() +
    geom_sf(data = map_obj, aes(fill = density)) +
    geom_point(data = obs_obj, aes(x = x, y = y))
  
}
