% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_ind.R
\name{simulate_ind}
\alias{simulate_ind}
\title{Simulate individuals}
\usage{
simulate_ind(map_obj, N, mean_group_size = 1, crs = 2154)
}
\arguments{
\item{map_obj}{sf dataframe. Sf map with a colum containg density informations}

\item{N}{numeric. The number of individuals desired in the area.}

\item{mean_group_size}{numeric. Mean size of groups to simulate. By default = 1.}

\item{crs}{numeric. Projection system. By default: 2154.}
}
\value{
The simulated individuals in a dataframe. The dataframe contains three columns, \code{x} longitude of the simulated individual, \code{y} latitude of the simulated individual and \code{size} the number of individuals simulated on the location (the possibility of simulate groupe will comming).
}
\description{
\code{simulate_ind} allows to simulate a approximate number of individuals with a inhomogenous Poisson point process according to the densities provided.
}
\details{
This function need a grid sf object with density assosciated to each square of the grid.
For each simulated individuals, a size is associated. Thus, it is possible to simulate group. Group are simulated using a Poisson distribution, with parameter \code{mean_group_size}. By default size is 1.
}
\examples{

library(ggplot2)
data("shape_courseulles")

# First, create a map with a gradient density from the North
map <- simulate_density(shape_obj = shape_courseulles,
                              grid_size = 1000,
                              density_type = "gradient",
                              gradient_direction = "N",
                              wavelength = 20000,
                              amplitude = 15)

# ------------------------------
# Example 1. Simulate the presence of 200 individuals 
# ------------------------------

ind <- simulate_ind(map_obj = map, N = 200)

# Plot
ggplot() +
  geom_sf(data = map, aes(fill = density)) +
  geom_point(data = ind, aes(x = x, y = y)) +
  theme(panel.background = element_rect(fill = "white"),
        panel.grid.major = element_line(colour = "#EDEDE9"))


# ------------------------------
# Example 2. Simulate the presence of 100 groups with a mean group size of 5 individuals 
# ------------------------------

ind <- simulate_ind(map_obj = map, mean_group_size = 5, N = 100)

# Plot
ggplot() +
  geom_sf(data = map) +
  geom_point(data = ind, aes(x = x, y = y, color = size)) +
  theme(panel.background = element_rect(fill = "white"),
        panel.grid.major = element_line(colour = "#EDEDE9"))
  
}
